+++ js/app/dashboard/redux/actions.js
@@ -10,12 +10,12 @@ import {
   createTaskListRequest,
   createTaskListSuccess,
   createTaskListFailure,
-  makeSelectTaskListItemsByUsername,
   enableUnassignedTourTasksDroppable,
   disableUnassignedTourTasksDroppable,
 } from '../../coopcycle-frontend-js/logistics/redux'
-import { selectNextWorkingDay, selectSelectedTasks } from './selectors'
+import { selectGroups, selectNextWorkingDay, selectSelectedTasks } from './selectors'
 import { selectUnassignedTours } from '../../../shared/src/logistics/redux/selectors'
+import { tourSelectors, taskSelectors } from './selectors'
 
 function createClient(dispatch) {
 
@@ -1216,7 +1216,8 @@ export function handleDragStart(result) {
   }
 }
 
-export function handleDragEnd(result) {
+// modifyTaskList is passed as argument so we can test this function thanks to dependency injection
+export function handleDragEnd(result, modifyTaskList) {
 
   return function(dispatch, getState) {
 
@@ -1258,18 +1259,44 @@ export function handleDragEnd(result) {
 
     const allTasks = selectAllTasks(getState())
 
+    // FIXME
+    // The tasks are dropped in the order they were selected
+    // Instead, we should respect the order of the unassigned tasks
+
+
+    // FIXME : if a tour or a group is selected, selectSelectedTasks yields [ undefined ] so we test > 1 no > 0
+    let selectedTasks = selectSelectedTasks(getState()).length > 1 ? selectSelectedTasks(getState()) : [_.find(allTasks, t => t['@id'] === result.draggableId)]
+
+    // we are moving a whole group or tour, override selectedTasks
+    if (result.draggableId.startsWith('group')) {
+      let groupId = result.draggableId.split(':')[1]
+      selectedTasks = selectGroups(getState()).find(g => g.id = groupId).tasks
+    }
+    else if (result.draggableId.startsWith('tour')) {
+      let tourId = result.draggableId.split(':')[1],
+      tour = tourSelectors.selectById(getState(), tourId)
+      selectedTasks = tour.itemIds.map(taskId => taskSelectors.selectById(getState(), taskId))
+    }
+    
+    // we want to move linked tasks together only when assigning and adding to a tour
+    // so we can keep fine grained control for reordering at will
+    if (source.droppableId !== destination.droppableId) {
+      selectedTasks =  withLinkedTasks(selectedTasks, allTasks, true)
+    }
+    
+    // handle drop in a tour
     if (destination.droppableId.startsWith('unassigned_tour:')) {
 
       const tours = selectUnassignedTours(getState())
       const tourId = destination.droppableId.replace('unassigned_tour:', '')
       const tour = tours.find(t => t['@id'] == tourId)
 
-      const newTourItems = [ ...tour.items ]
+      let newTourItems = [ ...tour.items ]
 
       // Drop new task into existing tour
       if (source.droppableId === 'unassigned') {
-        const task = _.find(allTasks, t => t['@id'] === result.draggableId)
-        newTourItems.splice(result.destination.index, 0, task)
+        Array.prototype.splice.apply(newTourItems,
+          Array.prototype.concat([ result.destination.index, 0 ], selectedTasks))
       }
 
       // Reorder tasks of existing tour
@@ -1283,70 +1310,27 @@ export function handleDragEnd(result) {
       return
     }
 
-    const taskLists = selectTaskLists(getState())
-    const selectedTasks = selectSelectedTasks(getState())
+    // handle drop in a tasklist
+    const tasksLists = selectTaskLists(getState())
 
     const username = destination.droppableId.replace('assigned:', '')
-    const taskList = _.find(taskLists, tl => tl.username === username)
-    const newTasks = [ ...taskList.items ]
+    const tasksList = _.find(tasksLists, tl => tl.username === username)
 
-    const selectTaskListItemsByUsername = makeSelectTaskListItemsByUsername()
-    const taskListItemsByUsername = selectTaskListItemsByUsername(getState(), { username })
+    let newTasksList = [...tasksList.items]
 
-    if (selectedTasks.length > 1) {
 
-      // FIXME Manage linked tasks
-      // FIXME
-      // The tasks are dropped in the order they were selected
-      // Instead, we should respect the order of the unassigned tasks
-
-      Array.prototype.splice.apply(newTasks,
-        Array.prototype.concat([ result.destination.index, 0 ], selectedTasks))
-
-    } else if (result.draggableId.startsWith('group:') || result.draggableId.startsWith('tour:')) {
-
-      const groupEl = document.querySelector(`[data-rbd-draggable-id="${result.draggableId}"]`)
-
-      const tasksFromGroup = Array
-        .from(groupEl.querySelectorAll('[data-task-id]'))
-        .map(el => _.find(allTasks, t => t['@id'] === el.getAttribute('data-task-id')))
-
-      Array.prototype.splice.apply(newTasks,
-        Array.prototype.concat([ result.destination.index, 0 ], tasksFromGroup))
-
-    } else {
-
-      // Reorder inside same list
-      if (source.droppableId === destination.droppableId) {
-        const [ removed ] = taskListItemsByUsername.splice(result.source.index, 1);
-        const newTaskListItemsByUsername = [ ...taskListItemsByUsername ]
-        newTaskListItemsByUsername.splice(result.destination.index, 0, removed)
-
-        // Flatten list
-        const flatArray = newTaskListItemsByUsername.reduce((items, item) => {
-          if (item['@type'] === 'Tour') {
-            item.items.forEach(t => items.push(t))
-          } else {
-            items.push(item)
-          }
-          return items
-        }, [])
-
-        newTasks.length = 0; // Clear the array
-        newTasks.push(...flatArray);
-
-      } else {
-        const task = _.find(allTasks, t => t['@id'] === result.draggableId)
-        if (task) {
-          const linkedTasks = withLinkedTasks(task, allTasks)
-          Array.prototype.splice.apply(newTasks,
-            Array.prototype.concat([ result.destination.index, 0 ], linkedTasks))
-        }
+    selectedTasks.forEach((task) => {
+      let taskIndex = newTasksList.findIndex((item) => item['@id'] === task['@id'])
+      // if the task was already in the tasklist, remove from its original place 
+      if ( taskIndex > -1) {
+        newTasksList.splice(taskIndex, 1)
       }
+    })
 
-    }
+    // insert all the selected tasks at the destination index
+    newTasksList.splice(result.destination.index, 0, ...selectedTasks)
 
-    dispatch(modifyTaskList(username, newTasks))
+    dispatch(modifyTaskList(username, newTasksList))
   }
 }
 
@@ -1675,9 +1659,10 @@ export function deleteTour(tour) {
 }
 
 export function removeTaskFromTour(tour, task) {
-
-  return function(dispatch) {
-    dispatch(modifyTour(tour, withoutTasks(tour.items, [ task ])))
+  return function(dispatch, getState) {
+    let state = getState()
+    let allTasks = selectAllTasks(state)
+    dispatch(modifyTour(tour, withoutTasks(tour.items, withLinkedTasks([ task ], allTasks))))
   }
 }