+++ src/Controller/AdminController.php
@@ -9,6 +9,7 @@ use AppBundle\Annotation\HideSoftDeleted;
 use AppBundle\Controller\Utils\AccessControlTrait;
 use AppBundle\Controller\Utils\AdminDashboardTrait;
 use AppBundle\Controller\Utils\DeliveryTrait;
+use AppBundle\Controller\Utils\InjectAuthTrait;
 use AppBundle\Controller\Utils\OrderTrait;
 use AppBundle\Controller\Utils\RestaurantTrait;
 use AppBundle\Controller\Utils\StoreTrait;
@@ -20,6 +21,7 @@ use AppBundle\Entity\User;
 use AppBundle\Entity\Delivery;
 use AppBundle\Entity\DeliveryForm;
 use AppBundle\Entity\DeliveryRepository;
+use AppBundle\Entity\Delivery\ImportQueue as DeliveryImportQueue;
 use AppBundle\Entity\Delivery\PricingRuleSet;
 use AppBundle\Entity\Hub;
 use AppBundle\Entity\BusinessAccount;
@@ -94,6 +96,9 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\ORM\Tools\Pagination\Paginator;
 use Doctrine\ORM\Query\Expr;
+use Hashids\Hashids;
+use League\Flysystem\Filesystem;
+use Lexik\Bundle\JWTAuthenticationBundle\Services\JWTTokenManagerInterface;
 use Nucleos\UserBundle\Model\UserManager as UserManagerInterface;
 use Nucleos\UserBundle\Util\TokenGenerator as TokenGeneratorInterface;
 use Nucleos\UserBundle\Util\Canonicalizer as CanonicalizerInterface;
@@ -129,6 +134,7 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use League\Bundle\OAuth2ServerBundle\Model\Client as OAuth2Client;
 use Twig\Environment as TwigEnvironment;
+use phpcent\Client as CentrifugoClient;
 
 class AdminController extends AbstractController
 {
@@ -141,6 +147,7 @@ class AdminController extends AbstractController
     use RestaurantTrait;
     use StoreTrait;
     use UserTrait;
+    use InjectAuthTrait;
 
     protected function getRestaurantRoutes()
     {
@@ -179,7 +186,8 @@ class AdminController extends AbstractController
         HttpClientInterface $browserlessClient,
         bool $optinExportUsersEnabled,
         CollectionFinderInterface $typesenseShopsFinder,
-        bool $adhocOrderEnabled
+        bool $adhocOrderEnabled,
+        protected JWTTokenManagerInterface $JWTTokenManager
     )
     {
         $this->orderRepository = $orderRepository;
@@ -204,15 +212,19 @@ class AdminController extends AbstractController
 
     protected function getOrderList(Request $request, $showCanceled = false)
     {
-        $qb = $this->orderRepository
-            ->createQueryBuilder('o');
-        $qb
-            ->andWhere('o.state != :state')
-            ->setParameter('state', OrderInterface::STATE_CART)
-            ->orderBy('LOWER(o.shippingTimeRange)', 'DESC')
-            ->setFirstResult(($request->query->getInt('p', 1) - 1) * self::ITEMS_PER_PAGE)
-            ->setMaxResults(self::ITEMS_PER_PAGE)
-            ;
+        if ($request->query->has('q')) {
+            $qb = $this->orderRepository->search($request->query->get('q'));
+        } else {
+            $qb = $this->orderRepository
+                ->createQueryBuilder('o');
+            $qb
+                ->andWhere('o.state != :state')
+                ->setParameter('state', OrderInterface::STATE_CART)
+                ->orderBy('LOWER(o.shippingTimeRange)', 'DESC')
+                ->setFirstResult(($request->query->getInt('p', 1) - 1) * self::ITEMS_PER_PAGE)
+                ->setMaxResults(self::ITEMS_PER_PAGE)
+                ;
+        }
 
         if (!$showCanceled) {
             $qb
@@ -238,17 +250,28 @@ class AdminController extends AbstractController
         OrderRepository $orderRepository
     )
     {
-        $results = $orderRepository->search($request->query->get('q'));
+        $qb = $orderRepository->search($request->query->get('q'));
+
+        $qb->setMaxResults(10);
+
+        $results = $qb->getQuery()->getResult();
 
         $data = [];
         foreach ($results as $order) {
-            $data[] = [
-                'id' => $order->getId(),
-                'name' => sprintf(
+
+            if (null !== $order->getCustomer()) {
+                $name = sprintf(
                     '%s (%s)',
                     $order->getNumber(),
                     $order->getCustomer()->getEmailCanonical()
-                ),
+                );
+            } else {
+                $name = $order->getNumber();
+            }
+
+            $data[] = [
+                'id' => $order->getId(),
+                'name' => $name,
                 'path' => $this->generateUrl('admin_order', ['id' => $order->getId()]),
             ];
         }
@@ -744,7 +767,11 @@ class AdminController extends AbstractController
         DeliveryManager $deliveryManager,
         OrderFactory $orderFactory,
         OrderManager $orderManager,
-        DeliveryRepository $deliveryRepository
+        DeliveryRepository $deliveryRepository,
+        Hashids $hashids8,
+        Filesystem $deliveryImportsFilesystem,
+        MessageBusInterface $messageBus,
+        CentrifugoClient $centrifugoClient
     )
     {
         $deliveryImportForm = $this->createForm(DeliveryImportType::class, null, [
@@ -753,21 +780,18 @@ class AdminController extends AbstractController
 
         $deliveryImportForm->handleRequest($request);
         if ($deliveryImportForm->isSubmitted() && $deliveryImportForm->isValid()) {
+
             $store = $deliveryImportForm->get('store')->getData();
 
-            return $this->handleDeliveryImportForStore($store, $deliveryImportForm, 'admin_deliveries',
-                $orderManager, $deliveryManager, $orderFactory);
-        } else if ($deliveryImportForm->isSubmitted() && !$deliveryImportForm->isValid()) {
-            if (count($deliveryImportForm->getData()) > 0) {
-                // This is the case when some rows have errors and some others were parsed successfuly and have to be persisted
-                $importedDeliveries = $this->persistImportedDeliveries($deliveryImportForm, $orderManager,
-                    $deliveryManager, $orderFactory);
-
-                $importedRowsMessage = $this->translator->trans('import.successful.rows', [
-                    '%count%' => count(array_keys($importedDeliveries)),
-                    '%rows%' => implode(", ", array_keys($importedDeliveries))
-                ]);
-            }
+            return $this->handleDeliveryImportForStore(
+                store: $store,
+                form: $deliveryImportForm,
+                messageBus: $messageBus,
+                entityManager: $this->entityManager,
+                filesystem: $deliveryImportsFilesystem,
+                hashids: $hashids8,
+                routeTo: 'admin_deliveries'
+            );
         }
 
         $dataExportForm = $this->createForm(DataExportType::class);
@@ -813,7 +837,7 @@ class AdminController extends AbstractController
             $deliveryRepository->searchWithSonic($qb, $filters['query'], $request->getLocale());
 
         } else {
-            if ($request->query->has('section') && is_callable([ $deliveryRepository, $request->query->get('section') ])) {
+            if ($request->query->has('section') && method_exists($deliveryRepository, $request->query->get('section'))) {
                 $qb = call_user_func([ $deliveryRepository, $request->query->get('section') ], $qb);
             } else {
                 $qb = $deliveryRepository->today($qb);
@@ -849,15 +873,25 @@ class AdminController extends AbstractController
             ]
         );
 
-        return $this->render('admin/deliveries.html.twig', [
+        $importQueues = $this->entityManager->getRepository(DeliveryImportQueue::class)
+            ->createQueryBuilder('diq')
+            ->andWhere('diq.createdAt >= :yesterday')
+            ->orderBy('diq.createdAt', 'DESC')
+            ->setParameter('yesterday', Carbon::yesterday())
+            ->getQuery()
+            ->getResult();
+
+        return $this->render('admin/deliveries.html.twig', $this->auth([
             'deliveries' => $deliveries,
             'filters' => $filters,
             'routes' => $this->getDeliveryRoutes(),
-            'imported_rows_message' => $importedRowsMessage ?? null,
             'stores' => $this->getDoctrine()->getRepository(Store::class)->findBy([], ['name' => 'ASC']),
             'delivery_import_form' => $deliveryImportForm->createView(),
             'delivery_export_form' => $dataExportForm->createView(),
-        ]);
+            'import_queues' => $importQueues,
+            'centrifugo_token' => $centrifugoClient->generateConnectionToken($this->getUser()->getUsername(), (time() + 3600)),
+            'centrifugo_channel' => sprintf('%s_events#%s', $this->getParameter('centrifugo_namespace'), $this->getUser()->getUsername()),
+        ]));
     }
 
     protected function getDeliveryRoutes()