+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.25.0] - 2024-01-16
+
+- Improve orders search
+- Improve dispatch dashboard drag'n'drop
+
+## [1.24.0] - 2024-01-12
+
+### Added
+
+- Fix 504 errors when uploading big delivery import files.
+- Show errors line by line in spreadsheet context.
+- Allow redownloading spreadsheet file with missing rows.
+
 ## [1.22.0] - 2023-12-20
 
 ### Added